testmessage: move TestMessageError into it's own module
